refactor(wsapi): split last files in smaller ones
