+++ event.go
@@ -1,6 +1,14 @@
 package gokord
 
 import (
+	"bytes"
+	"compress/zlib"
+	"encoding/json"
+	"io"
+	"time"
+
+	"github.com/gorilla/websocket"
+	"github.com/nyttikord/gokord/discord"
 	"github.com/nyttikord/gokord/event"
 	"github.com/nyttikord/gokord/guild"
 )
@@ -52,4 +60,109 @@ func (s *Session) onReady(r *event.Ready) {
 
 	// Store the ResumeGatewayURL within the Session struct.
 	s.resumeGatewayURL = r.ResumeGatewayURL
+}
+
+// onGatewayEvent is the "event handler" for all messages received on the Discord Gateway API websocket connection.
+func (s *Session) onGatewayEvent(messageType int, message []byte) (*discord.Event, error) {
+	var err error
+	var reader io.Reader
+	reader = bytes.NewBuffer(message)
+
+	// If this is a compressed message, uncompress it.
+	if messageType == websocket.BinaryMessage {
+		z, err2 := zlib.NewReader(reader)
+		if err2 != nil {
+			return nil, err2
+		}
+
+		defer func() {
+			err3 := z.Close()
+			if err3 != nil {
+				s.LogError(err3, "closing zlib")
+			}
+		}()
+
+		reader = z
+	}
+
+	var e *discord.Event
+	decoder := json.NewDecoder(reader)
+	if err = decoder.Decode(&e); err != nil {
+		return e, err
+	}
+
+	// handle special opcode
+	switch e.Operation {
+	case discord.GatewayOpCodeHello: // processed by Open()
+		return e, nil
+	case discord.GatewayOpCodeHeartbeat: // must respond with a heartbeat packet within 5 seconds
+		s.LogDebug("sending heartbeat in response to Op1")
+		return e, s.heartbeat(s.ws, s.sequence.Load())
+	case discord.GatewayOpCodeReconnect: // must immediately disconnect from gateway and reconnect to new gateway
+		s.LogInfo("closing and reconnecting in response to Op7")
+		err = s.CloseWithCode(websocket.CloseServiceRestart)
+		if err != nil {
+			s.LogError(err, "closing session connection, force closing")
+			s.ForceClose()
+		}
+		s.reconnect()
+		return e, nil
+	case discord.GatewayOpCodeInvalidSession: // must respond with an Identify packet
+		s.LogWarn("invalid session received, reconnecting")
+		err = s.CloseWithCode(websocket.CloseServiceRestart)
+		if err != nil {
+			s.LogError(err, "closing session connection, force closing")
+			s.ForceClose()
+		}
+
+		var resumable bool
+		if err = json.Unmarshal(e.RawData, &resumable); err != nil {
+			return e, err
+		}
+
+		if !resumable {
+			s.LogInfo("Gateway session is not resumable, discarding its information")
+			s.resumeGatewayURL = ""
+			s.sessionID = ""
+			s.sequence.Store(0)
+		}
+
+		s.reconnect()
+		return e, nil
+	case discord.GatewayOpCodeHeartbeatAck:
+		s.Lock()
+		s.LastHeartbeatAck = time.Now().UTC()
+		s.Unlock()
+		s.LogDebug("got heartbeat ACK")
+		return e, nil
+	}
+
+	// Do not try to Dispatch a non-Dispatch Message
+	if e.Operation != discord.GatewayOpCodeDispatch {
+		// But we probably should be doing something with them.
+		// TEMP
+		s.LogWarn("unknown Op: %d, Seq: %d, Type: %s, Data: %s, message: %s", e.Operation, e.Sequence, e.Type, string(e.RawData), string(message))
+		return e, nil
+	}
+
+	s.sequence.Store(e.Sequence)
+
+	if eh, ok := event.GetInterfaceProvider(e.Type); ok {
+		e.Struct = eh.New()
+
+		if err = json.Unmarshal(e.RawData, e.Struct); err != nil {
+			s.LogWarn("failed to unmarshal %s event, data: %s", e.Type, e.RawData)
+			// READY events are always emitted
+			if e.Type != event.ReadyType {
+				return nil, err
+			}
+		}
+
+		s.eventManager.EmitEvent(s, e.Type, e.Struct)
+	} else {
+		s.LogWarn("unknown event: Op: %d, Seq: %d, Type: %s, Data: %s", e.Operation, e.Sequence, e.Type, string(e.RawData))
+		s.eventManager.EmitEvent(s, event.EventType, e)
+	}
+
+	return e, nil
 }
