+++ session.go
@@ -90,7 +90,7 @@ type Session struct {
 	eventManager *event.Manager
 
 	// The websocket connection.
-	wsConn *websocket.Conn
+	ws *websocket.Conn
 
 	// When nil, the session is not listening.
 	listening chan any
@@ -140,28 +140,6 @@ type GatewayStatusUpdate struct {
 	Game   status.Activity `json:"game"`
 	Status string          `json:"status"`
 	AFK    bool            `json:"afk"`
-}
-
-// Identify is sent during initial handshake with the discord gateway.
-// https://discord.com/developers/docs/topics/gateway#identify
-type Identify struct {
-	Token          string              `json:"token"`
-	Properties     IdentifyProperties  `json:"properties"`
-	Compress       bool                `json:"compress"`
-	LargeThreshold int                 `json:"large_threshold"`
-	Shard          *[2]int             `json:"shard,omitempty"`
-	Presence       GatewayStatusUpdate `json:"presence,omitempty"`
-	Intents        discord.Intent      `json:"intents"`
-}
-
-// IdentifyProperties contains the "properties" portion of an Identify packet.
-// https://discord.com/developers/docs/topics/gateway#identify-identify-connection-properties
-type IdentifyProperties struct {
-	OS              string `json:"$os"`
-	Browser         string `json:"$browser"`
-	Device          string `json:"$device"`
-	Referer         string `json:"$referer"`
-	ReferringDomain string `json:"$referring_domain"`
 }
 
 // UserAPI returns an userapi.Requester to interact with the user package.
